fix: composer update no binary bug and failed version detection bug

+++ scripts/update-release.sh
@@ -12,15 +12,22 @@ VERSION_RANGE="$(cat ./build/VERSION_RANGE)"
 VERSION_MAJOR="$(cat ./build/VERSION_MAJOR)"
 IS_DEV_VERSION="$(echo "${VERSION_RANGE}" | grep -F -q -e '-dev' && echo '1' || echo '0')"
 PHAR_COMMIT="$(git rev-parse --verify HEAD)"
-IS_LTS_MODE="1"# Manual switch
+# Manual switch
+IS_LTS_MODE="1"
 
 git checkout gh-pages
 
 VERSION_ENV="dev"
 VERSION_TEXT="#dev"
+VERSION_TEXT_EXTRA=""
+
 if [ ${IS_DEV_VERSION} = "0" ]; then
     VERSION_ENV="latest"
     VERSION_TEXT="#normal"
+fi
+
+if [ ${IS_LTS_MODE} = "1" ]; then
+    VERSION_TEXT_EXTRA="#lts"
 fi
 
 updateLatestFolders() {
@@ -31,16 +38,17 @@ updateLatestFolders() {
     cp -rp "${SOURCE_FOLDER}/*" ./releases/${VERSION_ENV}/
     git add -A ./releases/${VERSION_ENV}/*
 
-    git commit -S -m "Update ${VERSION_ENV} release" -m "version: ${VERSION}" -m "version-env: ${VERSION_ENV}" -m "version-range: ${VERSION_RANGE}" -m "version-major: ${VERSION_MAJOR}" -m "${VERSION_TEXT}" -m "Commit: ${PHAR_COMMIT}"
+    git commit -S -m "Update ${VERSION_ENV} release" -m "version: ${VERSION}" -m "version-env: ${VERSION_ENV}" -m "version-range: ${VERSION_RANGE}" -m "version-major: ${VERSION_MAJOR}" -m "${VERSION_TEXT}" -m "${VERSION_TEXT_EXTRA}" -m "Commit: ${PHAR_COMMIT}"
 }
 
 doChangesForRelease() {
     SOURCE_FOLDER="$1"
     COMMIT_TEXT="Update version ${VERSION}"
-    if [ ! -d /releases/${VERSION}/ ]; then
+    if [ ! -d ./releases/${VERSION}/ ]; then
         COMMIT_TEXT="Release ${VERSION}"
-        # Create version folder
-        mkdir ./releases/${VERSION}/
+    fi
+    # Do not update major for LTS releases
+    if [ ! -d ./releases/${VERSION}/ ] && [ ${IS_LTS_MODE} = "0" ]; then
         if [ -L ./releases/${VERSION_MAJOR} ]; then
             # Unlink version env
             unlink ./releases/${VERSION_MAJOR}
@@ -50,10 +58,10 @@ doChangesForRelease() {
         ls -lah ./releases/${VERSION_MAJOR}
         git add -A "./releases/${VERSION_MAJOR}"
     fi
-    # Delete version folder
-    rm -rf ./releases/${VERSION}/*
+    # Delete version folder even if it does not exist
+    rm -rf ./releases/${VERSION}
     # Move source folder to version folder
-    mv "${SOURCE_FOLDER}" ./releases/${VERSION}/
+    mv "${SOURCE_FOLDER}" ./releases/${VERSION}
     # Add to GIT index
     git add -A ./releases/${VERSION}/
     if [ -L ./releases/${VERSION_RANGE} ]; then
@@ -66,12 +74,13 @@ doChangesForRelease() {
     # Add to GIT index
     git add -A "./releases/${VERSION_RANGE}"
     # Commit the changes
-    git commit -S -m "${COMMIT_TEXT}" -m "version: ${VERSION}" -m "version-range: ${VERSION_RANGE}" -m "${VERSION_TEXT}" -m "Commit: ${PHAR_COMMIT}"
+    git commit -S -m "${COMMIT_TEXT}" -m "version: ${VERSION}" -m "version-range: ${VERSION_RANGE}" -m "${VERSION_TEXT}" -m "${VERSION_TEXT_EXTRA}" -m "Commit: ${PHAR_COMMIT}"
 }
 
 doChangesForRelease "./build"
 
 if [ ${IS_LTS_MODE} = "0" ]; then
+    echo 'Updating the version named folder'
     updateLatestFolders "./releases/${VERSION}"
 else
     echo 'LTS mode, skipping update of version ENVs'
